% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvard-botanists.R
\name{bot_search}
\alias{bot_search}
\title{Harvard botanist index functions}
\usage{
bot_search(name = NULL, individual = FALSE, start = NULL, fuzzy = FALSE,
  remarks = NULL, speciality = NULL, country = NULL,
  is_collector = FALSE, is_author = FALSE, team = FALSE, error = stop,
  ...)
}
\description{
Harvard botanist index functions
}
\examples{
\dontrun{
# bot_search(name = "Asa Gray")
# bot_search(name = "A. Gray")
# bot_search(remarks = "harvard")
# bot_search(name = "Gray", fuzzy = TRUE)

## FIXME - this leads to a JSON parsing error because they give
##   bad JSON in some results, including this example
# bot_search(country = "China")
}
}
\keyword{internal}
