% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_check.R
\name{categorical_check}
\alias{categorical_check}
\title{Categorical check}
\usage{
categorical_check(data, column_name, valid_entries)
}
\arguments{
\item{data}{The dataset.}

\item{column_name}{The name of the column whose entries need to be checked.}

\item{valid_entries}{The expected entries for that column.}
}
\value{
a table of problematic rows that have values that do not match the expected values, displaying the problematic observation's ID, observed value in the given column, and error message.
}
\description{
Categorical check
}
\examples{
path <- system.file("extdata", "mousedata.xlsx", package = "scrubR")
joined_data <- join_sheets(path, "ID", c("Birth", "Body Weight", "Outcome"))
categorical_check(joined_data, "Treatment", c("Alive", "Dead"))
}
