% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Clean data.frame}
\usage{
clean_df(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Clean data.frame
}
\examples{
df <- sample_data_1

# Make a `clean_df` object
clean_df(df)
}

