% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector-funs.R
\name{collectors}
\alias{collectors}
\alias{coll_clean}
\title{Collector based cleaning}
\usage{
coll_clean(x, collector = NULL)
}
\arguments{
\item{x}{(data.frame) A data.frame}

\item{collector}{(character) Collector field to use. See Details.}

\item{drop}{(logical) Drop bad data points or not. Either way, we parse
out bade data points as an attribute you can access. Default: \code{TRUE}}
}
\value{
Returns a data.frame, with attributes
}
\description{
Collector based cleaning
}
\details{
Explanation of the functions:
\itemize{
\item coll_clean - Standardize collector names
}
}
\examples{
# df <- data.frame(
#   coll = c('K.F.P. Martius', 'C. F. P. Martius', 'C. F. P. von Martius'),
#   species = 'Poa annua',
#   lat = 1:3,
#   lon = 4:6,
#  stringsAsFactors = FALSE
# )

# Standardize names
# NROW(df)
# df <- dframe(df) \%>\% coll_clean()
# NROW(df)
# attr(df, "coll_clean")
}
\keyword{internal}
