% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coord-funs.R
\name{coords}
\alias{coord_impossible}
\alias{coord_incomplete}
\alias{coord_unlikely}
\alias{coords}
\title{Coordinate based cleaning}
\usage{
coord_incomplete(x, drop = TRUE)

coord_impossible(x, drop = TRUE)

coord_unlikely(x, drop = TRUE)
}
\arguments{
\item{x}{(data.frame) A data.frame}

\item{drop}{(logical) Drop bad data points or not. Either way, we parse
out bade data points as an attribute you can access. Default: \code{TRUE}}
}
\value{
Returns a data.frame, with attributes
}
\description{
Coordinate based cleaning
}
\examples{
df <- sample_data_1

# Remove impossible coordinates
NROW(df)
df <- clean_df(df) \%>\% coord_impossible()
NROW(df)
attr(df, "coord_impossible")

# Remove incomplete cases
NROW(df)
df_inc <- clean_df(df) \%>\% coord_incomplete()
NROW(df_inc)
# attr(df_inc, "coord_incomplete")

# Remove unlikely points
NROW(df)
df_unlikely <- clean_df(df) \%>\% coord_unlikely()
NROW(df_unlikely)
attr(df_unlikely, "coord_unlikely")
}

