% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup}
\alias{dedup}
\title{Deduplicate records}
\usage{
dedup(x, how = "one", tolerance = 0.9)
}
\arguments{
\item{x}{(data.frame) A data.frame}

\item{how}{(character) Drop bad data points or not. Either way, we parse
out bade data points as an attribute you can access. Default: \code{"one"}}

\item{tolerance}{(numeric) Score (0 to 1) at which to determine a match. You'll
want to inspect outputs closely to tweak this value based on your data, as
results can vary.}
}
\value{
Returns a data.frame, optionally with attributes
}
\description{
Deduplicate records
}
\examples{
df <- sample_data_1
smalldf <- df[1:20, ]
smalldf <- rbind(smalldf, smalldf[10,])
smalldf[21, "key"] <- 1088954555
NROW(smalldf)
dp <- clean_df(smalldf) \%>\% dedup()
NROW(dp)
attr(dp, "dups")

# Another example - more than one set of duplicates
df <- sample_data_1
twodups <- df[1:10, ]
twodups <- rbind(twodups, twodups[c(9, 10), ])
rownames(twodups) <- NULL
NROW(twodups)
dp <- clean_df(twodups) \%>\% dedup()
NROW(dp)
attr(dp, "dups")
}

