% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{dframe}
\alias{dframe}
\title{Compact data.frame}
\usage{
dframe(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Compact data.frame
}
\examples{
dframe(sample_data_1)
dframe(mtcars)
dframe(iris)
}

