% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregion.R
\name{ecoregion}
\alias{ecoregion}
\title{Filter points within ecoregions}
\usage{
ecoregion(x, dataset = "meow", ecoregion, lat = NULL, lon = NULL, drop = TRUE)
}
\arguments{
\item{x}{(data.frame) A data.frame}

\item{dataset}{(character) the dataset to use. one of: "meow" (Marine
Ecoregions of the World), "fao" (). See Details.}

\item{ecoregion}{(character) the ecoregion name. See Details.}

\item{lat, lon}{(character) Latitude and longitude column to use. See Details.}

\item{drop}{(logical) Drop bad data points or not. Either way, we parse out
bad data points as an attribute you can access. Default: \code{TRUE}}

\item{ignore.na}{(logical) To consider NA values as a bad point or not.
Default: \code{FALSE}}
}
\value{
Returns a data.frame, with attributes
}
\description{
Filter points within ecoregions
}
\details{
see \code{scrubr_cache} for managing the cache of data
}
\section{Datasets}{
 and ecoregions:
\itemize{
\item meow: https://opendata.arcgis.com/datasets/ed2be4cf8b7a451f84fd093c2e7660e3_0.geojson
\item fao:http://www.fao.org/geonetwork/srv/en/main.home?uuid=ac02a460-da52-11dc-9d70-0017f293bd28
http://www.fao.org/figis/geoserver/area/ows?service=WFS&request=GetFeature&version=1.0.0&typeName=area:FAO_AREAS&outputFormat=SHAPE-ZIP
http://www.fao.org/figis/geoserver/area/ows?service=WFS&request=GetFeature&version=1.0.0&typeName=area:FAO_AREAS&outputFormat=application/json
}
}

\section{Ecoregions}{

\itemize{
\item meow:
\item fao: asdfasdf
}
}

\examples{
## Marine Ecoregions of the World
wkt <- 'POLYGON((-119.8 12.2, -105.1 11.5, -106.1 21.6, -119.8 20.9, -119.8 12.2))'
res <- rgbif::occ_data(geometry = wkt, limit=300)$data
res2 <- sf::st_as_sf(res, coords = c("decimalLongitude", "decimalLatitude"))
res2 <- sf::st_set_crs(res2, 4326)
mapview::mapview(res2)
tmp <- ecoregion(dframe(res), dataset = "meow",
   ecoregion = "ECOREGION:Mexican Tropical Pacific")
tmp2 <- sf::st_as_sf(tmp, coords = c("decimalLongitude", "decimalLatitude"))
tmp2 <- sf::st_set_crs(tmp2, 4326)
mapview::mapview(tmp2)

## FAO
wkt <- 'POLYGON((72.2 38.5,-173.6 38.5,-173.6 -41.5,72.2 -41.5,72.2 38.5))'
manta_ray <- rgbif::name_backbone("Mobula alfredi")$usageKey
res <- rgbif::occ_data(manta_ray, geometry = wkt, limit=300, hasCoordinate = TRUE)
dat <- sf::st_as_sf(res$data, coords = c("decimalLongitude", "decimalLatitude"))
dat <- sf::st_set_crs(dat, 4326)
mapview::mapview(dat)
tmp <- ecoregion(dframe(res$data), dataset = "fao", ecoregion = "OCEAN:Indian")
library(dplyr)
tmp <- filter(tmp, !is.na(decimalLongitude))
tmp2 <- sf::st_as_sf(tmp, coords = c("decimalLongitude", "decimalLatitude"))
tmp2 <- sf::st_set_crs(tmp2, 4326)
mapview::mapview(tmp2)
}
