% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_sheets.R
\name{join_sheets}
\alias{join_sheets}
\title{Join excel sheets
Join excel sheets and check for mismatches in the key and flag the mismatched entries}
\usage{
join_sheets(file_path, shared_column, selected_sheets = NULL)
}
\arguments{
\item{file_path}{The file path for the xlsx file.}

\item{shared_column}{The column to use as key for join.}

\item{selected_sheets}{The sheets to be joined.}
}
\value{
a joined data set and a if a mismatch is present, a table with the mismatched keys and the corresponding sheet name
}
\description{
Join excel sheets
Join excel sheets and check for mismatches in the key and flag the mismatched entries
}
\examples{
path <- system.file("extdata", "mousedata.xlsx", package = "scrubR")
joined_data <- join_sheets(path, "ID", c("Birth", "Body Weight", "Outcome"))
}
