% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change.R
\name{plot_change}
\alias{plot_change}
\title{Plot change over time
Create a line graph of different body weights recorded at different data points. Color and shape are mapped to the treatment or placebo group. If one entry (e.g. body weight) has dropped by more than 20\% of its previous value, it is marked as "Flagged" when the user hovers over the data point. Interactive so that the user can hover over data points and see the ID, value drop percentage, and whether the entry was flagged.}
\usage{
plot_change(data, id_col, var_cols, color, cat)
}
\arguments{
\item{data}{The dataset.}

\item{id_col}{The ID column has the individual observations.}

\item{var_cols}{The column containing the variable that is recorded over time and plotted.}

\item{color}{The variable mapped to the color of the graph.}

\item{cat}{The variable mapped to the shape of the geom.}
}
\value{
an interactive plot of the change of multiple discrete entries over time, color coded by category and flagged for significant (+20\%) drops in value
}
\description{
Plot change over time
Create a line graph of different body weights recorded at different data points. Color and shape are mapped to the treatment or placebo group. If one entry (e.g. body weight) has dropped by more than 20\% of its previous value, it is marked as "Flagged" when the user hovers over the data point. Interactive so that the user can hover over data points and see the ID, value drop percentage, and whether the entry was flagged.
}
\examples{
path <- system.file("extdata", "mousedata.xlsx", package = "scrubR")
joined_data <- join_sheets(path, "ID", c("Birth", "Body Weight", "Outcome"))
plot_change(joined_data, "ID",
c("Body Weight 1", "Body Weight 2", "Body Weight 3"),
"Treatment", "Treatment")
}
