% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_check.R
\name{quantitative_check}
\alias{quantitative_check}
\title{Quantitative check
For quantitative variables, check and flag whether values for a specified column are not numerical}
\usage{
quantitative_check(data, column_name)
}
\arguments{
\item{data}{The dataset.}

\item{column_name}{The column whose values need to be checked.

Return a table of problematic rows that have non-numerical values for a specified column, displaying the problematic observation's ID, observed value in the given column, and error message}
}
\description{
Quantitative check
For quantitative variables, check and flag whether values for a specified column are not numerical
}
\examples{
path <- system.file("extdata", "mousedata.xlsx", package = "scrubR")
joined_data <- join_sheets(path, "ID", c("Birth", "Body Weight", "Outcome"))
quantitative_check_table <- quantitative_check(joined_data, "Body Weight 3")
}
