% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_1}
\alias{sample_data_1}
\title{Sample data.frame 1}
\format{A data frame with 1500 rows and 5 variables:
\describe{
  \item{name}{taxonomic name}
  \item{longitude}{longitude, decimal degree}
  \item{latitude}{latitude, decimal degree}
  \item{date}{date, date the occurrence was recorded}
  \item{key}{occurrence key}
}

Data originally collected from GBIF}
\description{
Sample data.frame 1
}
\keyword{datasets}
\keyword{internal}
