% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_2}
\alias{sample_data_2}
\title{Sample data.frame number 2}
\format{A data frame with 1500 rows and 7 variables:
\describe{
\item{name}{taxonomic name}
\item{longitude}{longitude, decimal degree}
\item{latitude}{latitude, decimal degree}
\item{key}{occurrence key}
\item{year}{year of occurrence}
\item{month}{month of occurrence}
\item{day}{day of occurrence}
}

Data originally collected from GBIF}
\description{
Sample data.frame number 2
}
\keyword{datasets}
\keyword{internal}
