% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_3}
\alias{sample_data_3}
\title{Sample data.frame number 3}
\format{A data frame with 200 rows and 5 variables:
\describe{
\item{name}{taxonomic name}
\item{decimalLatitude}{latitude, decimal degree}
\item{decimalLongitude}{longitude, decimal degree}
\item{key}{occurrence key}
\item{eventDate}{date, date the occurrence was recorded}
}

Data originally collected from GBIF}
\description{
Sample data.frame number 3
}
\keyword{datasets}
\keyword{internal}
