% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_4}
\alias{sample_data_4}
\title{Sample data.frame number 4}
\format{A data frame with 100 rows and 6 variables:
\describe{
  \item{name}{taxonomic name}
  \item{decimalLatitude}{latitude, decimal degree}
  \item{decimalLongitude}{longitude, decimal degree}
  \item{key}{occurrence key}
  \item{eventDate}{date, date the occurrence was recorded}
  \item{country}{country of occurrence record}
}

Data originally collected from GBIF}
\description{
Sample data.frame number 4
}
\keyword{datasets}
\keyword{internal}
