% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_5}
\alias{sample_data_5}
\title{Sample data.frame number 5}
\format{A data frame with 39 rows and 5 variables:
\describe{
\item{name}{taxonomic name}
\item{longitude}{longitude, decimal degree}
\item{latitude}{latitude, decimal degree}
\item{date}{date, date the occurrence was recorded}
\item{key}{GBIF occurrence key}
}

Data originally collected from GBIF}
\description{
Sample data.frame number 5
}
\keyword{datasets}
\keyword{internal}
