% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_6}
\alias{sample_data_6}
\title{Sample data.frame number 6}
\format{A data frame with 50 rows and 5 variables:
\describe{
\item{name}{taxonomic name}
\item{key}{GBIF occurrence key}
\item{decimalLatitude}{latitude, decimal degree}
\item{decimalLongitude}{longitude, decimal degree}
\item{coordinateUncertaintyInMeters}{Uncertainity, the point-radius representation of the location}
}

Data originally collected from GBIF}
\description{
Sample data.frame number 6
}
\keyword{datasets}
\keyword{internal}
