% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\docType{data}
\name{sample_data_7}
\alias{sample_data_7}
\title{Sample data.frame number 7}
\format{
A data frame with 50 rows and 91 variables

Data originally collected from GBIF
}
\description{
Sample data.frame number 7
}
\keyword{datasets}
\keyword{internal}
