% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr-package.R
\docType{package}
\name{scrubr-package}
\alias{scrubr}
\alias{scrubr-package}
\title{scrubr - Clean biological occurrence data}
\description{
scrubr - Clean biological occurrence data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

