% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-funs.R
\name{taxonomy}
\alias{tax_}
\alias{taxonomy}
\title{Taxonomy based cleaning}
\usage{
tax_(x, drop = TRUE)
}
\arguments{
\item{x}{(data.frame) A data.frame}

\item{drop}{(logical) Drop bad data points or not. Either way, we parse
out bade data points as an attribute you can access. Default: \code{TRUE}}
}
\value{
Returns a data.frame, with attributes
}
\description{
Taxonomy based cleaning
}
\examples{
df <- sample_data_1

# Remove impossible coordinates
NROW(df)
df <- clean_df(df) \%>\% coord_impossible()
NROW(df)
attr(df, "coord_impossible")
}

