% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coord-funs.R
\name{coords}
\alias{coord_impossible}
\alias{coord_incomplete}
\alias{coords}
\title{Coordinate based cleaning}
\usage{
coord_incomplete(x)

coord_impossible(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
Returns a data.frame, with attributes
}
\description{
Coordinate based cleaning
}
\examples{
\dontrun{
df <- sample_data_1

# Make a `clean_df` object
clean_df(df)

# Remove impossible coordinates
NROW(df)
df <- clean_df(df) \%>\% coord_impossible()
NROW(df)
attr(df, "coord_impossible")

# Remove incomplete cases
NROW(df)
df_inc <- clean_df(df) \%>\% coord_incomplete()
NROW(df_inc)
attr(df_inc, "coord_incomplete")
}
}

