% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubr_datasets.R
\name{scrubr_datasets}
\alias{scrubr_datasets}
\title{scrubr datasets}
\description{
\itemize{
\item \code{\link{sample_data_1}}
\item \code{\link{sample_data_2}}
\item \code{\link{sample_data_3}}
\item \code{\link{sample_data_4}}
\item \code{\link{sample_data_6}}
}
}
\keyword{datasets}
